Move double tests to constraints.c/h.
